restructure: shared 폴더 src 하위로 이동
