+++ src/shared/avatar/Avatar.tsx
