+++ src/shared/badges/CategoryBadge.tsx
