+++ src/shared/badges/DistrictBadge.tsx
